Add files via upload
The pre-alpha 1.0 version of Osmosis Writer. My first actual bit of software. As of this stage, Osmosis Writer can:
create, load, and save novel projects (.osm files)
create chapters within novel projects (.osmc files)
allow users to rearrange their chapters via dragging and dropping the left hand tabs
export the user's novel to a .doc file, with in-built formatting

improvements to make for this current version:
OW crashes if the user exits out of the "open file" dialog
Loading an existing project when a current project is open will put both on the screen (and in the work tabs)
The console prints out so much garbage. Must clean up garbage.
lots of unused code to reimplement or remove
and many, many more